+++ components/SwipperEmpresas/SwipperEmpresas.module.css
@@ -1,0 +1,113 @@
+.itemList {
+    border: 1px solid #eaeaea;
+    padding: 10px;
+    margin-bottom: 30px;
+    overflow: hidden; 
+    background-color: #fff;
+  }
+  
+  .item:hover {
+    border: 1px solid #000;
+  }
+  
+  .ancor {
+    text-decoration: none;
+  }
+  
+  .figure {
+    overflow: hidden;
+    position: relative;
+  }
+  
+  .photo {
+    width: 100%;
+    object-fit: cover;
+  }
+  
+  .info {
+    padding: 15px;
+  }
+  
+  .name {
+    font-size: 13px;
+    line-height: 21px;
+    font-weight: 400;
+    min-height: 50px;
+    color: #3C3C3B;
+  }
+  
+  .blockAndWeight {
+    display: block;
+    font-weight: 500;
+  }
+  
+  .btn {
+    font-family: 'Inter', sans-serif;
+    font-weight: 500;
+    margin-top: 10px; 
+    font-size: .7rem;
+    background-color: transparent;
+    border-color: #000000;
+    color: #212529;
+    padding: .5rem 1.5rem;
+    line-height: 1.5;
+    border-radius: .3rem;
+  }
+  
+  .btn:hover {
+    color: #212529;
+    background-color: #e2e6ea;
+    border-color: #dae0e5;
+  }
+  
+  .swiperButtonNext, .swiperButtonPrev {
+    color: #fff;
+    background: rgba(255, 255, 255, 0.7);
+    padding: 10px;
+    width: auto;
+    height: auto;
+    border-radius: 10px;
+    backdrop-filter: blur(5px);
+  }
+
+  .swiperButtonNext:active, .swiperButtonPrev:active {
+    background: rgb(255,202,5);
+  }
+
+  .swiperButtonNext:hover {
+    color: #000;
+    animation: moveDirecctionRight 0.4s infinite alternate;
+    /* left: 0; */
+  } 
+  .swiperButtonPrev:hover {
+    color: #000;
+    animation: moveDirecctionLeft 0.4s infinite alternate;
+    /* right: 0; */
+  }
+
+  @keyframes moveDirecctionLeft {
+    0% {
+      left: 0;
+    }
+    100%{
+      left: 10px;
+    }
+  }
+
+  @keyframes moveDirecctionRight {
+    0% {
+      right: 0;
+    }
+    100%{
+      right: 10px;
+    }
+  }
+
+  .photoPreview {
+    width: 58px;
+    height: 58px;
+    object-fit: cover;
+    position: absolute;
+    top: 45%;
+    margin: 10px;
+  }
