+++ locales/en/home.json
@@ -1,3 +1,7 @@
 {
-  "welcome-to": "Welcome to"
+  "welcome-to": "Welcome to",
+  "products" : "Products",
+  "design-objects-for-your-spaces" : "Design objects for your spaces",
+  "professionals": "Professionals",
+  "new-design-and-construction-professionals": "New design and construction professionals"
 }
