+++ locales/es/home.json
@@ -1,3 +1,7 @@
 {
-  "welcome-to": "Bienvenido a"
+  "welcome-to": "Bienvenido a",
+  "products": "Productos",
+  "design-objects-for-your-spaces": "Objetos de diseño para tus espacios",
+  "professionals": "Profesionales",
+  "new-design-and-construction-professionals": "Nuevos profesionales de diseño y la construcción"
 }
