+++ styles/Home.module.css
@@ -124,3 +124,21 @@
 .inlineCode::after {
   content: '`';
 }
+
+.itemsTitle {
+  margin: auto;
+  font-size: 15px;
+  font-weight: 600;
+}
+
+.itemsSmallTitle {
+  font-size: 12px;
+  font-weight: 400;
+  display: block;
+  margin-top: 5px;
+}
+
+.infoHead {
+  text-align: center;
+  margin-bottom: 30px;
+}
